requirements.txt
# Core ML Libraries
torch>=2.1.0
transformers>=4.35.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Dataset and Processing
datasets>=2.14.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# AWS Integration
boto3>=1.28.0
botocore>=1.31.0

# Logging and Monitoring
loguru>=0.7.0
tensorboard>=2.14.0
wandb>=0.15.0  # Optional: Weights & Biases integration

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0
numpy>=1.24.0
scipy>=1.11.0
